Extracts Tika util
