+++ stages/processing/tika/pom.xml
@@ -16,13 +16,13 @@
 		<dependency>
 			<groupId>com.findwise.hydra</groupId>
 			<artifactId>hydra-api</artifactId>
-			<version>0.6.0-SNAPSHOT</version>
+			<version>${project.parent.version}</version>
 			<type>jar</type>
 		</dependency>
 		<dependency>
-			<groupId>org.apache.tika</groupId>
-			<artifactId>tika-parsers</artifactId>
-			<version>1.1</version>
+			<groupId>com.findwise.hydra</groupId>
+			<artifactId>hydra-tika-utils</artifactId>
+			<version>${project.parent.version}</version>
 		</dependency>
 		<dependency>
 			<groupId>junit</groupId>
@@ -61,11 +61,6 @@
 			<plugin>
 				<groupId>com.ning.maven.plugins</groupId>
 				<artifactId>maven-duplicate-finder-plugin</artifactId>
-				<configuration>
-					<ignoredResources>
-						<ignoredResource>META-INF/BCKEY.*</ignoredResource>
-					</ignoredResources>
-				</configuration>
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
@@ -81,25 +76,7 @@
 						<transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
 							<mainClass>com.findwise.hydra.stage.GroupStarter</mainClass>
 						</transformer>
-
-						<!-- This merges the service resources, the important one in our case 
-							being META-INF/services/org.apache.tika.parser.Parser -->
-						<transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer" />
 					</transformers>
-
-					<!-- We need to remove any jar signatures since they won't match ours. 
-						Otherwise we get java.lang.SecurityException: Invalid signature file digest 
-						for Manifest main attributes -->
-					<filters>
-						<filter>
-							<artifact>*:*</artifact>
-							<excludes>
-								<exclude>META-INF/*.SF</exclude>
-								<exclude>META-INF/*.DSA</exclude>
-								<exclude>META-INF/*.RSA</exclude>
-							</excludes>
-						</filter>
-					</filters>
 				</configuration>
 				<executions>
 					<execution>
