+++ stages/processing/tika/src/main/java/com/findwise/hydra/stage/tika/TikaStage.java
@@ -14,7 +14,7 @@ import com.findwise.hydra.local.LocalDocument;
 import com.findwise.hydra.stage.AbstractProcessStage;
 import com.findwise.hydra.stage.Parameter;
 import com.findwise.hydra.stage.Stage;
-import com.findwise.hydra.stage.tika.utils.TikaUtils;
+import com.findwise.utils.tika.TikaUtils;
 
 /**
  * @author jwestberg
