+++ stages/utils/tika/src/main/java/com/findwise/utils/tika/TikaUtils.java
@@ -1,4 +1,4 @@
-package com.findwise.hydra.stage.tika.utils;
+package com.findwise.utils.tika;
 
 import java.io.IOException;
 import java.io.InputStream;
